% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotting_fns.R
\name{barPlot}
\alias{barPlot}
\title{Create a bar plot}
\usage{
barPlot(statsObj)
}
\arguments{
\item{statsObj}{A DAMstats obj (created by \code{\link{calcStats}})}
}
\value{
A ggplot2 plot object.
}
\description{
Create a bar plot from calculated statistics (plots data means by attribute).
Error bars represent the standard error of the mean. Plot is automatically
faced if using multiple timepoints. Do not use this function for datasets
with a large number of timepoints (use \code{\link{linePlot}} instead). All
ggplot2 functions will work on the resulting plot.
}
\examples{
sleep <- dropDead(DAM_DD)
sleep <- calcSleep(sleep)
sleep <- toInterval(sleep, 24, units = "hours", aggregateBy = "average")
sleep <- toAvgDay(sleep)
stat <- calcStats(sleep, "genotype")
barPlot(stat)
}

