% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/by_drop_attribute.R
\name{byAttribute}
\alias{byAttribute}
\title{Subset data by an attribute}
\usage{
byAttribute(obj, string, col)
}
\arguments{
\item{obj}{A valid DAM object (created by \code{\link{newExperiment}}).}

\item{string}{Attribute values to include. Can be a vector.}

\item{col}{Which column of attribute values should be used.}
}
\value{
Returns a DAM S4 object.
}
\description{
A more generic function to subset your data by whichever column you  want.
Even works on vectors, so you can subset by multiple values of an attribute
at a time. To remove data by an attribute, see \code{\link{dropAttribute}}.
}
\examples{
listAttribVals(DAM_DD, "genotype")
returnedDAM <- byAttribute(DAM_DD, "experimental")
listAttribVals(returnedDAM, "genotype")

listAttribVals(DAM_DD, "genotype")
returnedDAM <- byAttribute(DAM_DD, c("control A", "control B"))
listAttribVals(returnedDAM, "genotype")
}
\seealso{
\code{\link{dropAttribute}}
}

