% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcActivityIndex.R
\name{calcActivityIndex}
\alias{calcActivityIndex}
\title{Calculate activity index}
\usage{
calcActivityIndex(obj)
}
\arguments{
\item{obj}{A valid DAM S4 object}
}
\value{
Returns a vector of the activity indexes for each fly
}
\description{
The activity index of a fly is the amount of activity counts it had during a
day / the total time it was awake.
}
\examples{
calcActivityIndex(dropDead(DAM_DD))
}

