% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAnovaN.R
\name{calcANOVA}
\alias{calcANOVA}
\title{Calculate statistics for a dataset}
\usage{
calcANOVA(obj, attribute, ..., vector)
}
\arguments{
\item{obj}{A valid DAM object (created by \code{\link{newExperiment}}}

\item{attribute}{Which attribute stats should be calculated against.}
}
\value{
Returns
}
\description{
A 1-way ANOVA is performed across an attribute on data with a single 
timepoint. A 2-way ANOVA is performed across an attribute on data with 
multiple timepoints (this will take quite awhile if done on data with 
an extreme number of timepoints). Note that a 1-way ANOVA is statistically
identical to a two-sample t-test when there are only two sample groups.
}
\examples{
activity <- toInterval(DAM_DD, 12, units = "hours", aggregateBy = "sum")
activity <- toAvgDay(activity)
statsAct <- calcStats(activity, "genotype")
}

