% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{calcAttribMeans}
\alias{calcAttribMeans}
\title{Calculate average values by attribute}
\usage{
calcAttribMeans(obj, attribute, ..., vector)
}
\arguments{
\item{obj}{A valid DAM S4 object (created by \code{\link{newExperiment}})}

\item{attribute}{Which attribute you would like to use to analyze your data
by, e.g. "genotype"}

\item{vector}{A vector of values to use for mean calculation instead of using
  those in actual DAM object obj.}
}
\value{
A matrix of values with the average and standard error for each
  attribute category
}
\description{
This function is used to calculate the average values
(sleep/activity/whatever) of your data by an attribute. For instance, this
could be used to calculate the average sleep values for each genotype.
}
\examples{
sleep <- dropDead(DAM_DD)
sleep <- calcSleep(sleep)
calcAttribMeans(sleep, "genotype")
}

