% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSleepBouts.R
\name{calcBouts}
\alias{calcBouts}
\title{Calculate bout number and duration for a single animal}
\usage{
calcBouts(vector)
}
\arguments{
\item{vector}{This is a 1D vector pertaining to a single fly's sleep data.}
}
\value{
Returns sleep bout and duration in terms of number of indices where
  the fly was asleep.
}
\description{
Calculate bout number and duration for a single animal
}

