% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSleepBouts.R
\name{calcMeanBout}
\alias{calcMeanBout}
\title{Calculate mean duration of each sleep bout}
\usage{
calcMeanBout(obj)
}
\arguments{
\item{obj}{A valid DAM S4 object where \code{\link{calcSleep}} has already
been performed.}
}
\value{
A named vector with the duration of each bout in minutes.
}
\description{
This method calculates the mean duration of each sleep bout in minutes.
}
\examples{
sleep <- calcSleep(dropDead(DAM_DD))
calcMeanBout(sleep)
}

