% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calcSleepBouts.R
\name{calcNumBouts}
\alias{calcNumBouts}
\title{Calculate mean number of sleep bouts per day}
\usage{
calcNumBouts(obj)
}
\arguments{
\item{obj}{A valid DAM S4 object where \code{\link{calcSleep}} has already
  been performed.}
}
\value{
A named vector of the number of sleep bouts/length of your experiment in days
}
\description{
This function calculates the mean number of sleep bouts per day for each fly.
}
\examples{
sleep <- calcSleep(dropDead(DAM_DD))
calcNumBouts(sleep)
}

