% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calcSleep.R
\name{calcSleep}
\alias{calcSleep}
\title{Calculate sleep}
\usage{
calcSleep(obj)
}
\arguments{
\item{obj}{A valid DAM object (created by \code{\link{newExperiment}})}
}
\value{
A DAM S4 object.
}
\description{
Calculates the amount of sleep in a dataset. This function is recommended to
be run on datasets with 5 minute intervals between measurements. Data at
"faster" rates will be coerced to 5 minutes.
}
\examples{
calcSleep(DAM_DD)
}

