% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calcStats.R
\name{calcStats}
\alias{calcStats}
\title{Calculate statistics}
\usage{
calcStats(obj, attribute)
}
\arguments{
\item{obj}{A valid DAM object (created by \code{\link{newExperiment}}}

\item{attribute}{Which attribute stats should be calculated against.}
}
\value{
A DAMstats S4 Object
}
\description{
Iterate through possible variations of an attribute in your dataset,
calculating averages/standard error of the mean/stats for each. Returns a
DAMstats object.
}
\details{
A 1-way ANOVA is performed across an attribute on data with a single
timepoint. A 2-way ANOVA is performed across an attribute on data with
multiple timepoints (for performance reasons, this is not done on data with
an extreme number of timepoints).
}
\examples{
activity <- toInterval(DAM_DD, 1, units = "hours", aggregateBy = "sum")
statsAct <- calcStats(activity, "genotype")
}

