% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/survivalTime.R
\name{calcSurvivalTime}
\alias{calcSurvivalTime}
\title{Calculate time of death}
\usage{
calcSurvivalTime(obj, ..., endThresh = 4)
}
\arguments{
\item{obj}{A valid DAM S4 object.}

\item{...}{}

\item{endThresh}{An optional parameter specifying how long (in hours) an
  animal must remain motionless before it can be considered "dead." Does not
  affect the index of death.}
}
\value{
Returns the index at which animals in a dataset died. Returns NAs for
  animals that remained alive at experiment conclusion.
}
\description{
This method determines at which point an animal died during an experiment.
This is computed by finding the last time period in which an animal moved. If
an animal continued moving (and was alive) until the end of an experiment,
this function returns NA for that animal.
}
\examples{
calcSurvivalTime(DAM_DD)
}

