% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catExperiments.R
\name{catExperiments}
\alias{catExperiments}
\title{Concatenate experiments}
\usage{
catExperiments(obj_vector)
}
\arguments{
\item{obj_vector}{A vector of datasets you wish to concatenate. Time 
intervals for each dataset must be the same.}
}
\value{
A DAM S4 object.
}
\description{
Use this function to mash experiments together. There is no restriction to 
doing this by timestamp, etc. - so be careful what you combine. The vial 
indices in the second experiment are relabeled to start where the first 
experiment stops. All metadata is maintained throughout the operation. Longer
experiments are coerced down to the size of the smallest experiment.
}
\examples{
biggerDAM <- catExperiments(c(DAM_DD, DAM_DD))
}

