% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DAM.class.R
\docType{class}
\name{DAM-class}
\alias{DAM-class}
\title{DAM experiment object}
\description{
This is an S4 object designed to hold the output produced by Trikinetics'
Drosophila Activity Monitor in its native format as well as its metadata.
}
\section{Slots}{

\describe{
\item{\code{data}}{A dataframe containing the output of a Drosophila Activity Monitor
experiment.}

\item{\code{sample_info}}{A dataframe that contains the metadata for an activity
  monitor experiment, with one entry for every vial present.}
}}

