% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropAnimals.R
\name{dropAnimals}
\alias{dropAnimals}
\title{Remove an animal from the experiment}
\usage{
dropAnimals(obj, vialNumbers)
}
\arguments{
\item{obj}{A valid DAM S4 object}

\item{vialNumbers}{Which vial numbers to remove from a dataset. Can be a 
vector.}
}
\value{
A DAM S4 object.
}
\description{
This function removes a specified animal from a dataset. Can operate on 
vectors, removing multiple animals in a single operation. To remove animals 
by their metadata, such as "remove all females", use
\code{\link{dropAttribute}} instead.
}
\examples{
returnedDAM <- dropAnimals(DAM_DD, 12)

returnedDAM <- dropAnimals(DAM_DD, c(1:3))
}
\seealso{
\code{\link{dropAttribute}}
}

