% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by_drop_attribute.R
\name{dropAttribute}
\alias{dropAttribute}
\title{Remove data by an attribute}
\usage{
dropAttribute(obj, string, col)
}
\arguments{
\item{obj}{A valid DAM object (created by \code{\link{newExperiment}}).}

\item{string}{Attribute values to include. Can be a vector.}

\item{col}{Which column of attribute values should be used.}
}
\value{
Returns a DAM S4 object.
}
\description{
Equivalent to \code{\link{byAttribute}}, but drops data by a value instead. Works on vectors.
}
\examples{
listAttribVals(DAM_DD, "genotype")
returnedDAM <- byAttribute(DAM_DD, "experimental")
listAttribVals(returnedDAM, "genotype")

listAttribVals(DAM_DD, "genotype")
returnedDAM <- byAttribute(DAM_DD, c("control A", "control B"))
listAttribVals(returnedDAM, "genotype")

}
\seealso{
\code{\link{byAttribute}}
}

