% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/exportDataset.R
\name{exportDataset}
\alias{exportDataset}
\title{Export a DAM experiment}
\usage{
exportDataset(obj, filename)
}
\arguments{
\item{obj}{A valid DAM S4 object (created by \code{\link{newExperiment}})}

\item{filename}{The filename MINUS THE EXTENSION that you wish to export your
  data to. Note that file extensions will automatically added.}
}
\value{
Writes two files to your current path: one with an experiment's raw
  data, and another with its metadata.
}
\description{
This is a convenience method to export your data for use with other software.
Creates two files, a .txt file with the raw data for your dataset, and a .csv
file with all of its metadata. The .txt file will be in the original
TriKinetics format. NOTE: if you have performed any sleep calculations on a
dataset, it will confuse other software packages. Export only activity counts
if you wish to use a dataset in other DAM applications.
}
\examples{
\dontrun{
exportDataset(DAM_DD, "baseFilename")
}
}

