% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findLightInterval.R
\name{findLightInterval}
\alias{findLightInterval}
\title{Calculate the lighting interval}
\usage{
findLightInterval(DAM)
}
\arguments{
\item{DAM}{A valid DAM S4 object}
}
\value{
The lighting interval (in hours)
}
\description{
Automatically calculates the lighting interval of a dataset.
}

