% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getInterval.R
\name{getInterval}
\alias{getInterval}
\title{Determine measurement interval}
\usage{
getInterval(obj)
}
\arguments{
\item{obj}{A valid DAM S4 object.}
}
\value{
Time interval between measurements in seconds.
}
\description{
This method calculates the time difference between measurements in your
dataset.
}
\examples{
getInterval(DAM_DD)
}

