% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getLightChanges.R
\name{getLightChanges}
\alias{getLightChanges}
\title{Find light changes in a dataset}
\usage{
getLightChanges(DAM)
}
\arguments{
\item{DAM}{A valid DAM S4 object}
}
\value{
A vector of indices where light status changed.
}
\description{
This is a utility method that returns indices where a Drosophila Activity
Monitor detects a change in lighting status.
}

