% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSleepBouts.R
\name{getSleepBouts}
\alias{getSleepBouts}
\title{Calculate duration and number of sleep bouts}
\usage{
getSleepBouts(obj)
}
\arguments{
\item{obj}{A valid DAM S4 object (created by \code{\link{newExperiment}})}
}
\value{
Returns a list with the duration of each sleep bout (in minutes).
}
\description{
Returns a list with the duration of each sleep bout for each fly.
}

