% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVals.R
\name{getVals}
\alias{getVals}
\title{Retrieve values from a DAM dataframe as a matrix.}
\usage{
getVals(DAMobject)
}
\arguments{
\item{DAMobject}{This is a dataframe corresponding to a DAM objects data}
}
\value{
a matrix of values
}
\description{
Retrieve values from a DAM dataframe as a matrix.
}

