% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSleep.R
\name{isAsleep}
\alias{isAsleep}
\title{Calculate sleep values}
\usage{
isAsleep(count)
}
\arguments{
\item{count}{A raw number of counts}
}
\value{
Whether or not a fly was asleep during that interval.
}
\description{
Changes raw counts to sleep yes/no over that interval. Do not use time
intervals under 5 minutes.
}
\examples{
isAsleep(5)
isAsleep(NA)
}

