% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{linePlot}
\alias{linePlot}
\title{Create a line plot of an experiment}
\usage{
linePlot(obj, attribute, ...)
}
\arguments{
\item{obj}{A valid DAM S4 object}

\item{attribute}{The attribute of the data you wish to examine (like "genotype")}
}
\value{
Returns a ggplot2 plot object
}
\description{
Creates a line plot of the data. Useful for datasets with a lot of 
timepoints. Line datapoints represent the mean for that attribute. The 
lightly colored regions represent the standard error of the mean. Grey/white
areas of the background represent time periods where the lights were on/off.
Most useful for datasets with a large number of timepoints.
}
\examples{
sleep <- dropDead(DAM_DD)
sleep <- calcSleep(sleep)
sleep <- toInterval(sleep, 1, units = "hours", aggregateBy = "average")
linePlot(sleep, "genotype") 
}
\seealso{
For datasets with only a few timepoints, use \code{\link{barPlot}} 
  instead.
}

