% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_fns_new.R
\name{linePlot2}
\alias{linePlot2}
\title{Create a line plot of an experiment}
\usage{
linePlot2(obj, attribute)
}
\arguments{
\item{obj}{}

\item{attribute}{}
}
\value{
Returns a ggplot2 plot object
}
\description{
Create a line plot of an experiment
}
\examples{
DAM_DD
}

