% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/attributeMethods.R
\name{listAttribVals}
\alias{listAttribVals}
\title{Get attribute values}
\usage{
listAttribVals(obj, attribute)
}
\arguments{
\item{obj}{A valid DAM S4 object}

\item{attribute}{A string that matches one of the column names from the
  original sample info .csv file}
}
\value{
A string of attribute values remaining in a dataset.
}
\description{
A convenience method to show the possible values of an attribute. These are
the values from a column in the original .csv metadata file.
}
\examples{
listAttribVals(DAM_DD, "genotype")
}
\seealso{
\code{\link{listAttributes}}
}

