% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/listToDataframe.R
\name{listToDataframe}
\alias{listToDataframe}
\title{Convert a list to a dataframe}
\usage{
listToDataframe(list)
}
\arguments{
\item{list}{A list}
}
\value{
Returns a dataframe
}
\description{
Converts a list to a datframe. Each element of the list gets its own column.
If elements of the list are of different lengths, NA's will be added to the
end of each column.
}

