% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAM.class.R
\name{newExperiment}
\alias{newExperiment}
\title{Create a DAM experiment}
\usage{
newExperiment(dataFile = NULL, infoFile = NULL)
}
\arguments{
\item{dataFile}{A string containing the path to a valid DAM output .txt file.}

\item{infoFile}{A string containing the path to a .csv (comma-separated 
values) file that you have created for the experiment. The only 
requirements for this file is that it must contain a 1-line header (these 
are the attribute names), and the first column must contain the vial
numbers used in the experiment (typically 1 through 32).}
}
\value{
A DAM S4 object.
}
\description{
Create a DAM experiment
}
\examples{
\dontrun{
DAM <- newExperiment(dataFile = "/path/to/DAM_output.txt", infoFile = "/path/to/metadata.csv")
}
}

