% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributeMethods.R
\name{orderAttribVals}
\alias{orderAttribVals}
\title{Change attribute order}
\usage{
orderAttribVals(obj, attribute, order)
}
\arguments{
\item{obj}{A valid DAM S4 object}

\item{attribute}{A string that matches one of the column names from the 
original sample info .csv file}

\item{order}{A vector with the new order of values for the attribute. Can be
either a character vector with the}
}
\value{
Returns a DAM object with the attribute of concern reordered
}
\description{
This is a convenience function that allows easy reordering of attribute 
values for plotting (does not affect data analysis). Bars in the wrong order?
Use this function.
}
\examples{
# reorder by number
dat <- data(DAM_DD)
listAttribVals(DAM_DD, "genotype")
dat <- orderAttribVals(DAM_DD, "genotype", c(3, 2, 1))
listAttribVals(DAM_DD, "genotype")

# reorder by values
dat <- data(DAM_DD)
listAttribVals(DAM_DD, "genotype")
dat <- orderAttribVals(DAM_DD, "genotype", c("experimental", "control B", "control A"))
listAttribVals(DAM_DD, "genotype")
}
\seealso{
To see the current attribute order, use \code{\link{listAttribVals}}
}

