% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentToMins.R
\name{percentToMins}
\alias{percentToMins}
\title{Convert sleep percentages to minutes}
\usage{
percentToMins(obj)
}
\arguments{
\item{obj}{A valid DAM S4 object (\code{\link{calcSleep}} must have already 
been performed)}
}
\value{
A DAM dataset with sleep values in minutes.
}
\description{
This method converts a dataset with sleep percentages to a dataset with sleep
measured in minutes. DONT RUN THIS ON DATA WHERE YOU HAVE NOT CALCULATED
SLEEP.
}
\examples{
sleep <- calcSleep(DAM_DD)
sleep <- percentToMins(sleep)
}

