% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotting_fns.R
\name{prepPlot}
\alias{prepPlot}
\title{Reshape data in preparation for plotting}
\usage{
prepPlot(statsObj)
}
\arguments{
\item{statsObj}{}
}
\value{
A list of meta and plotting data.
}
\description{
Performs the necessary data reshaping before a plot can be made
}

