% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_fns_new.R
\name{prepPlot2}
\alias{prepPlot2}
\title{Prepare data for plotting}
\usage{
prepPlot2(obj, attribute)
}
\arguments{
\item{obj}{}

\item{attribute}{}
}
\value{
A matrix of values with the average and sem for each attribute
  category
}
\description{
Prepare data for plotting
}
\examples{
DAM_DD
}

