% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/subsetTime.R
\name{subsetTime}
\alias{subsetTime}
\title{Subsetting data by time}
\usage{
subsetTime(obj, startTime, expDuration, units)
}
\arguments{
\item{obj}{A valid DAM object (created by \code{\link{newExperiment}}).}

\item{startTime}{The starting point to include.}

\item{expDuration}{The duration of time you wish to include.}

\item{units}{The units of time you are using. Can be one of the following:
  "days", "hours", "minutes", or "seconds."}
}
\value{
Returns a DAM S4 object.
}
\description{
This function subsets your data to correspond to a certain time interval. Use
this function to subset out a particular portion of an experiment (to clip
off unused data from beginnning and end).
}
\examples{
returnedDAM <- subset(DAM, startTime = 1, expDuration = 3, units = "days")
}

