% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/survivalTime.R
\name{survivalTime}
\alias{survivalTime}
\title{Computes the index at which a fly died in the data}
\usage{
survivalTime(vector, threshold)
}
\arguments{
\item{vector}{A column of the dataset to operate on.}

\item{threshold}{The threshold defines a base number of counts, below which
  could be considered spurious counts. Designed to prevent a random count
  here and there from affecting the time-of-death measurement.}
}
\value{
Time-of-death (NA if fly was alive)
}
\description{
The algorithm that computes time-of-death for flies in a DAM experiment.
}

