% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/toAvgDay.R
\name{toAvgDay}
\alias{toAvgDay}
\title{Compute average day}
\usage{
toAvgDay(DAM, ..., incomplete.rm = TRUE)
}
\arguments{
\item{DAM}{A valid DAM S4 experiment}

\item{...}{}

\item{incomplete.rm}{A boolean value specifying whether incomplete days at
  the end of the dataset should not be included in the calculation. Default:
  TRUE}
}
\value{
Returns a DAM S4 experiment containing one day's worth of averaged
  measurements.
}
\description{
Averages the data from multiple days to create an averaged "typical day."
Data from each individual animal is handled separately.
}
\examples{
toAvgDay(DAM_DD)
}

