% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toInterval.R
\name{toInterval}
\alias{toInterval}
\title{Collapse a dataset to a target time interval}
\usage{
toInterval(obj, target, units, aggregateBy)
}
\arguments{
\item{obj}{A valid DAM experiment object}

\item{target}{What time interval should the resulting dataset be?}

\item{units}{The units used by the "target" parameter. Can be one of the 
following: "seconds", "minutes", "hours", "days"}

\item{aggregateBy}{When aggregating multiple timepoints' worth of data, 
should data be averaged or summed? Can be one of the following: "sum", 
"average"}
}
\value{
Returns a DAM S4 object, with measurments at the specified time 
  interval.
}
\description{
Use this function to change the time between measurements of a DAM dataset 
(can only be used to increase the interval between measurements). When 
compressing data to the new set of timepoints, data is aggreagated by either
mean or sum.
}
\examples{
getInterval(DAM_DD)
returnedDAM <- toInterval(DAM_DD, 1, units = "hours", aggregateBy = "sum")
getInterval(returnedDAM)
}

