% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toSeconds.R
\name{toSeconds}
\alias{toSeconds}
\title{Convert a time to seconds}
\usage{
toSeconds(time, units = c("seconds", "minutes", "hours", "days"))
}
\arguments{
\item{units}{}
}
\value{
time in seconds
}
\description{
A utility function used to converts "units" args in other functions to
seconds.
}

