% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toTidy.R
\name{toTidy}
\alias{toTidy}
\title{Convert a dataset to TidyR format}
\usage{
toTidy(obj)
}
\arguments{
\item{obj}{A DAM S4 experiment}
}
\value{
Returns a dataframe in TidyR format.
}
\description{
This method converts a data and all its associated metadata to the TidyR 
format. See 
\link{http://cran.r-project.org/web/packages/tidyr/vignettes/tidy-data.html}
for details.
}
\examples{
tidy <- toTidy(DAM_DD)
}

