% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_fns_new.R
\name{barPlot}
\alias{barPlot}
\title{Create a bar plot of an experiment}
\usage{
barPlot(obj, attribute, ..., vector)
}
\arguments{
\item{obj}{A valid DAM S4 object}

\item{attribute}{The attribute of the data you wish to examine (like "genotype")}

\item{vector}{A vector of values, with one for each animal in the experiment}
}
\value{
Returns a ggplot2 plot object
}
\description{
Create a bar plot from calculated statistics (plots data means by attribute).
Error bars represent the standard error of the mean. Plot is automatically
faced if using multiple timepoints. Do not use this function for datasets
with a large number of timepoints (use \code{\link{linePlot}} instead). All
ggplot2 functions will work on the resulting plot.
}
\examples{
sleep <- dropDead(DAM_DD)
sleep <- calcSleep(sleep)
sleep <- toInterval(sleep, 24, units = "hours", aggregateBy = "average")
sleep <- toAvgDay(sleep)
barPlot(sleep, "genotype")

sleep <- dropDead(DAM_DD)
sleep <- calcSleep(sleep)
bouts <- calcMeanBout(sleep)
barPlot(sleep, "genotype", bouts)
}

