% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotting_fns.R
\name{linePlot}
\alias{linePlot}
\title{Create a line plot of the data}
\usage{
linePlot(statsObj)
}
\arguments{
\item{statsObj}{A DAMstats obj (created by \code{\link{calcStats}})}
}
\value{
Returns a standard ggplot2 object.
}
\description{
Creates a line plot of the data. Useful for datasets with a lot of
timepoints. Line datapoints represent the mean for that attribute. The
lightly colored regions represent the standard error of the mean. Grey/white
areas of the background represent time periods where the lights were on/off.
Most useful for datasets with a large number of timepoints.
}
\examples{
sleep <- dropDead(DAM_DD)
sleep <- calcSleep(sleep)
sleep <- toInterval(sleep, 1, units = "hours", aggregateBy = "average")
stat <- calcStats(sleep, "genotype")
barPlot(stat)
}
\seealso{
For datasets with only a few timepoints, use \code{\link{barPlot}}
  instead.
}

