% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/whichChanged.R
\name{whichChanged}
\alias{whichChanged}
\title{Find changes in a vector}
\usage{
whichChanged(vector, delta)
}
\arguments{
\item{vector}{}

\item{delta}{}
}
\value{
indices which changed more than delta
}
\description{
Find indices of which values in a vector differed by more than a certain
delta
}

